app/requirements.txt
pycryptodome
pathlib
requests
